Missing requst response Dto added

+++ server/src/application/use-cases/company/delete-company-workplace-picture.use-case.ts
@@ -1,15 +1,19 @@
 import { ICompanyWorkplacePicturesRepository } from '../../../domain/interfaces/repositories/company/ICompanyWorkplacePicturesRepository';
-import { NotFoundError } from '../../../domain/errors/errors';
-import { IDeleteCompanyWorkplacePictureUseCase } from '../../../domain/interfaces/use-cases/ICompanyUseCases';
+import { NotFoundError, AuthorizationError } from '../../../domain/errors/errors';
+import { IDeleteCompanyWorkplacePictureUseCase } from '../../../domain/interfaces/use-cases/company/IDeleteCompanyWorkplacePictureUseCase';
 
 export class DeleteCompanyWorkplacePictureUseCase implements IDeleteCompanyWorkplacePictureUseCase {
   constructor(private readonly _companyWorkplacePicturesRepository: ICompanyWorkplacePicturesRepository) {}
 
-  async execute(pictureId: string): Promise<void> {
-    const deleted = await this._companyWorkplacePicturesRepository.delete(pictureId);
-    if (!deleted) {
+  async execute(companyId: string, pictureId: string): Promise<void> {
+    const existingPicture = await this._companyWorkplacePicturesRepository.findById(pictureId);
+    if (!existingPicture) {
       throw new NotFoundError(`Company workplace picture with ID ${pictureId} not found`);
     }
+    if (existingPicture.companyId !== companyId) {
+      throw new AuthorizationError('Not authorized to delete this workplace picture');
+    }
+    await this._companyWorkplacePicturesRepository.delete(pictureId);
   }
 }
